fix: runtime cast crash, stale fallback version, and const correctness
- summary_screen: `? 0 :` → `? 0.0 :` in two places — prevents a
  runtime `TypeError` (`int` cast to `double`) when reserved > balance
- summary_screen: remove the now-redundant `as double` cast on the same
  expression; type is already inferred as `double`
- section_screen: add missing `const _Divider()` constructor so both
  call-sites can be `const _Divider()`
- language_provider: `Locale('pt')` → `const Locale('pt')` on the
  field initialiser
- about_screen: stale hardcoded fallback `'2.0.2'` → `''` — the real
  version is loaded from PackageInfo almost immediately
- models/reserved_amount: promote constructor to `const`

+++ lib/screens/summary_screen.dart
@@ -59,13 +59,13 @@ class _SummaryScreenState extends State<SummaryScreen> {
 
       final balance = transactions.fold<double>(0, (sum, t) => sum + (t.isCredit ? t.amount : -t.amount),);
 
-      final availableAmount = (balance - totalReserved) < 0 ? 0 : balance - totalReserved;
+      final availableAmount = (balance - totalReserved) < 0 ? 0.0 : balance - totalReserved;
 
       setState(() {
         _allTransactions = transactions;
         _allReservedAmounts = reservedAmounts;
         _totalReserved = totalReserved;
-        _availableAmount = availableAmount as double;
+        _availableAmount = availableAmount;
         _selectedMonths = _getAvailableMonths();
         _isLoading = false;
       });
@@ -99,7 +99,7 @@ class _SummaryScreenState extends State<SummaryScreen> {
     final totalExpenses = transactions.where((t) => !t.isCredit).fold<double>(0, (sum, t) => sum + t.amount);
     final balance = totalIncome - totalExpenses;
 
-    final availableAmount = (balance - _totalReserved) < 0 ? 0 : balance - _totalReserved;
+    final availableAmount = (balance - _totalReserved) < 0 ? 0.0 : balance - _totalReserved;
 
     return {
       'totalIncome': totalIncome,
